Randomise some charges

+++ electric_flutter/lib/Charge.dart
@@ -3,8 +3,8 @@ import 'dart:math';
 /// Electric charge particle.
 ///
 /// @author Moshe Waisberg
-class Charge extends Point<int> {
-  Charge(int x, int y, this.size) : super(x, y);
+class Charge extends Point<double> {
+  Charge(double x, double y, this.size) : super(x, y);
 
   final double size;
